Re-organize mixer code
