+++ e2etest/get_stats_test.go
@@ -20,66 +20,16 @@ import (
 	"io/ioutil"
 	"path"
 	"runtime"
-	"strings"
 	"testing"
 
 	pb "github.com/datacommonsorg/mixer/proto"
-	"github.com/datacommonsorg/mixer/server"
 	"github.com/google/go-cmp/cmp"
-	"google.golang.org/grpc"
+	"google.golang.org/protobuf/testing/protocmp"
 )
-
-// This test runs agains staging staging bt and bq dataset.
-// This is billed to GCP project "datcom-ci"
-// It needs Application Default Credentials to run locally or need to
-// provide service account credential when running on GCP.
-const (
-	btProject   = "google.com:datcom-store-dev"
-	btInstance  = "prophet-cache"
-	bqProject   = "datcom-ci"
-	branchCache = false
-)
-
-func Setup(ctx context.Context) (pb.MixerClient, error) {
-	_, filename, _, _ := runtime.Caller(0)
-	btTable, _ := ioutil.ReadFile(
-		path.Join(path.Dir(filename), "../deployment/staging_bt_table.txt"))
-	bqTable, _ := ioutil.ReadFile(
-		path.Join(path.Dir(filename), "../deployment/staging_bq_table.txt"))
-	// Use mapping template before we need to test SPARQL query.
-	schemaPath := path.Join(path.Dir(filename), "../deployment/mapping")
-
-	s, err := server.NewServer(
-		"localhost:0",
-		strings.TrimSpace(string(bqTable)),
-		strings.TrimSpace(string(btTable)),
-		btProject,
-		btInstance,
-		bqProject,
-		schemaPath,
-		branchCache,
-	)
-	if err != nil {
-		return nil, err
-	}
-	// Start mixer at localhost:0
-	go s.Srv.Serve(s.Lis)
-
-	// Create mixer client
-	conn, err := grpc.Dial(
-		s.Addr,
-		grpc.WithInsecure(),
-		grpc.WithDefaultCallOptions(grpc.MaxCallRecvMsgSize(100000000 /* 100M */)))
-	if err != nil {
-		return nil, err
-	}
-	client := pb.NewMixerClient(conn)
-	return client, nil
-}
 
 func TestGetStats(t *testing.T) {
 	ctx := context.Background()
-	client, err := Setup(ctx)
+	client, err := setup()
 	if err != nil {
 		t.Fatalf("Failed to set up mixer and client")
 	}
@@ -112,7 +62,6 @@ func TestGetStats(t *testing.T) {
 			false,
 		},
 	} {
-
 		resp, err := client.GetStats(ctx, &pb.GetStatsRequest{
 			StatsVar: c.statsVar,
 			Place:    c.place,
@@ -146,11 +95,10 @@ func TestGetStats(t *testing.T) {
 				}
 			}
 		} else {
-			if diff := cmp.Diff(result, expected); diff != "" {
+			if diff := cmp.Diff(result, expected, protocmp.Transform()); diff != "" {
 				t.Errorf("payload got diff: %v", diff)
 				continue
 			}
 		}
-
 	}
 }
