+++ e2etest/get_triples_test.go
@@ -23,13 +23,13 @@ import (
 	"testing"
 
 	pb "github.com/datacommonsorg/mixer/proto"
-	"github.com/datacommonsorg/mixer/store"
+	"github.com/datacommonsorg/mixer/server"
 	"github.com/google/go-cmp/cmp"
 )
 
 func TestGetTriples(t *testing.T) {
 	ctx := context.Background()
-	client, err := Setup(ctx)
+	client, err := setup()
 	if err != nil {
 		t.Fatalf("Failed to set up mixer and client")
 	}
@@ -77,7 +77,7 @@ func TestGetTriples(t *testing.T) {
 			"",
 			false,
 			5,
-			[]int{42, 22},
+			[]int{43, 27},
 		},
 	} {
 		req := &pb.GetTriplesRequest{Dcids: c.dcids}
@@ -89,7 +89,7 @@ func TestGetTriples(t *testing.T) {
 			t.Errorf("could not GetTriples: %s", err)
 			continue
 		}
-		var result map[string][]*store.Triple
+		var result map[string][]*server.Triple
 		err = json.Unmarshal([]byte(resp.GetPayload()), &result)
 		if err != nil {
 			t.Errorf("Can not Unmarshal payload")
@@ -107,7 +107,7 @@ func TestGetTriples(t *testing.T) {
 			continue
 		}
 
-		var expected map[string][]*store.Triple
+		var expected map[string][]*server.Triple
 		file, _ := ioutil.ReadFile(path.Join(goldenPath, c.goldenFile))
 		err = json.Unmarshal(file, &expected)
 		if err != nil {
