+++ go.mod
@@ -3,23 +3,34 @@ module github.com/datacommonsorg/mixer
 go 1.12
 
 require (
-	cloud.google.com/go v0.59.0 // indirect
-	cloud.google.com/go/bigquery v1.8.0
-	cloud.google.com/go/bigtable v1.3.0
+	cloud.google.com/go/bigquery v1.9.0
+	cloud.google.com/go/bigtable v1.4.0
 	cloud.google.com/go/pubsub v1.4.0
 	cloud.google.com/go/storage v1.10.0
+	github.com/cosiner/argv v0.1.0 // indirect
 	github.com/deckarep/golang-set v1.7.1
-	github.com/go-test/deep v1.0.3
+	github.com/go-delve/delve v1.4.1 // indirect
+	github.com/go-test/deep v1.0.6
 	github.com/golang/protobuf v1.4.2
 	github.com/google/go-cmp v0.5.0
+	github.com/mattn/go-colorable v0.1.7 // indirect
+	github.com/mattn/go-runewidth v0.0.9 // indirect
+	github.com/peterh/liner v1.2.0 // indirect
+	github.com/spf13/cobra v1.0.0 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
 	go.opencensus.io v0.22.4 // indirect
+	go.starlark.net v0.0.0-20200619143648-50ca820fafb9 // indirect
+	golang.org/x/arch v0.0.0-20200511175325-f7c78586839d // indirect
 	golang.org/x/net v0.0.0-20200625001655-4c5254603344 // indirect
 	golang.org/x/sync v0.0.0-20200625203802-6e8e738ad208
 	golang.org/x/sys v0.0.0-20200625212154-ddb9806d33ae // indirect
 	golang.org/x/text v0.3.3 // indirect
-	golang.org/x/tools v0.0.0-20200626171337-aa94e735be7f // indirect
+	golang.org/x/tools v0.0.0-20200702044944-0cc1aa72b347 // indirect
 	google.golang.org/api v0.28.0
-	google.golang.org/genproto v0.0.0-20200626011028-ee7919e894b5
+	google.golang.org/genproto v0.0.0-20200702021140-07506425bd67
 	google.golang.org/grpc v1.30.0
+	google.golang.org/grpc/cmd/protoc-gen-go-grpc v0.0.0-20200630190442-3de8449f8555 // indirect
 	google.golang.org/protobuf v1.25.0
+	gopkg.in/check.v1 v1.0.0-20190902080502-41f04d3bba15 // indirect
+	gopkg.in/yaml.v2 v2.3.0 // indirect
 )
