+++ proto/mixer.proto
@@ -14,6 +14,8 @@
 
 syntax = "proto3";
 
+// Use relative go package so the generated file is in the current folder.
+option go_package = "./proto";
 package datacommons;
 
 import "google/api/annotations.proto";
