--- requirements.txt
@@ -1,3 +1,0 @@
-sphinx
-sphinxcontrib-openapi
-sphinx_rtd_theme
