+++ server/property_label_test.go
@@ -1,4 +1,18 @@
-package store
+// Copyright 2020 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package server
 
 import (
 	"context"
@@ -61,20 +75,22 @@ func TestGetPropertyLabels(t *testing.T) {
 		Payload: string(wantPayloadRaw),
 	}
 
-	store, err := NewTestBtStore(ctx, data)
+	btTable, err := setupBigtable(ctx, data)
 	if err != nil {
 		t.Fatalf("NewTestBtStore() = %+v", err)
 	}
 
-	var out pb.GetPropertyLabelsResponse
-	if err := store.GetPropertyLabels(ctx,
+	s := NewServer(nil, btTable, nil, nil, "")
+
+	got, err := s.GetPropertyLabels(ctx,
 		&pb.GetPropertyLabelsRequest{
 			Dcids: []string{"geoId/06", "bio/tiger"},
-		}, &out); err != nil {
+		})
+	if err != nil {
 		t.Fatalf("GetPropertyLabels() = %+v", err)
 	}
 
-	if got := &out; !cmp.Equal(got, want, protocmp.Transform()) {
+	if !cmp.Equal(got, want, protocmp.Transform()) {
 		t.Errorf("GetPropertyLabels() = %+v, want %+v", got, want)
 	}
 }
