+++ server/stats_test.go
@@ -1,4 +1,18 @@
-package store
+// Copyright 2020 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package server
 
 import (
 	"encoding/json"
@@ -146,8 +160,16 @@ const populationJSON = `{
 func TestTriplesToStatsVar(t *testing.T) {
 	var covidStatsVarTriples TriplesCache
 	var populationStatsVarTriples TriplesCache
-	json.Unmarshal([]byte(covidJSON), &covidStatsVarTriples)
-	json.Unmarshal([]byte(populationJSON), &populationStatsVarTriples)
+	err := json.Unmarshal([]byte(covidJSON), &covidStatsVarTriples)
+	if err != nil {
+		t.Errorf("Unmarshal got err %v", err)
+		return
+	}
+	err = json.Unmarshal([]byte(populationJSON), &populationStatsVarTriples)
+	if err != nil {
+		t.Errorf("Unmarshal got err %v", err)
+		return
+	}
 	for _, c := range []struct {
 		triples      *TriplesCache
 		wantStatsVar *StatisticalVariable
