--- store/places_test.go
@@ -1,8 +1,0 @@
-package store
-
-import (
-	"testing"
-)
-
-func TestGetPlacesIn(t *testing.T) {
-}
